chore(generators): generate fake data
Related work item: GH-21
