+++ .gitignore
@@ -34,6 +34,5 @@ buildcache/
 # Other
 node_modules
 dist
-/firebase/seed
 dist-ssr
 *.local
