+++ firebase/seed/auth_export/accounts.json
@@ -1,0 +1,1 @@
+{ "kind": "identitytoolkit#DownloadAccountResponse", "users": [] }
