+++ firebase/seed/firebase-export-metadata.json
@@ -1,0 +1,16 @@
+{
+  "version": "11.24.1",
+  "firestore": {
+    "version": "1.16.0",
+    "path": "firestore_export",
+    "metadata_file": "firestore_export/firestore_export.overall_export_metadata"
+  },
+  "auth": {
+    "version": "11.24.1",
+    "path": "auth_export"
+  },
+  "storage": {
+    "version": "11.24.1",
+    "path": "storage_export"
+  }
+}
