+++ firebase/seed/storage_export/buckets.json
@@ -1,0 +1,7 @@
+{
+  "buckets": [
+    {
+      "id": "warzywniax.appspot.com"
+    }
+  ]
+}
