+++ tsconfig.json
@@ -30,6 +30,7 @@
     "forceConsistentCasingInFileNames": true,
     "jsx": "react-jsx",
     "resolveJsonModule": true,
+    "esModuleInterop": true,
     "baseUrl": "src"
   }
 }
